refactor: extract router
